test(MailService): migrate tests to TypeScript
